Introduced data classes outside the NativeProxy
