+++ compiler/src/main/java/io/github/landerlyoung/jenny/FieldIdDeclaration.kt
@@ -1,0 +1,9 @@
+package io.github.landerlyoung.jenny
+
+import javax.lang.model.element.VariableElement
+
+
+data class FieldIdDeclaration(
+    val helper: HandyHelper,
+    val listOfFields: List<VariableElement>
+)
