+++ compiler/src/main/java/io/github/landerlyoung/jenny/MethodIdDeclaration.kt
@@ -1,0 +1,6 @@
+package io.github.landerlyoung.jenny
+
+data class MethodIdDeclaration(
+    val helper: HandyHelper,
+    val listOfMethods: List<MethodOverloadResolver.MethodRecord>
+)
