+++ sample-android/templates/fields_ids_declarations.kte
@@ -1,4 +1,4 @@
-@import io.github.landerlyoung.jenny.NativeProxyGenerator.FieldIdDeclaration
+@import io.github.landerlyoung.jenny.FieldIdDeclaration
 @import io.github.landerlyoung.jenny.HandyHelper
 @import javax.lang.model.element.VariableElement
