+++ sample-android/templates/fields_ids_initialisations.kte
@@ -1,6 +1,7 @@
-@import io.github.landerlyoung.jenny.NativeProxyGenerator.FieldIdDeclaration
+@import io.github.landerlyoung.jenny.FieldIdDeclaration
 @import io.github.landerlyoung.jenny.HandyHelper
 @import javax.lang.model.element.VariableElement
+@import javax.lang.model.element.Modifier
 
 @param  fieldIdDeclaration:FieldIdDeclaration
