+++ sample-android/templates/methods_ids_declarations.kte
@@ -1,4 +1,4 @@
-@import io.github.landerlyoung.jenny.NativeProxyGenerator.MethodIdDeclaration
+@import io.github.landerlyoung.jenny.MethodIdDeclaration
 @import io.github.landerlyoung.jenny.HandyHelper
 @import io.github.landerlyoung.jenny.MethodOverloadResolver.MethodRecord
